automation-assistant/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
anthropic>=0.18.0
pyyaml>=6.0
aiohttp>=3.9.0
websockets>=12.0
pydantic>=2.5.0
